+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# [1.2.0-alpha.4]
+
+- (QA) Removed setup popup if was already initialized at some point
+- (QA) Improved performance when using airways
+- Added SID/STAR/Approach selection to each airport
+- (QA) Added STAR holdings
+
 # [1.2.0-alpha.3]
 
 - (QA) Added waypoints to holdings